Implementing supplier price List
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C27285

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Business/SupplierZoneManager.cs
@@ -16,8 +16,8 @@ namespace TOne.WhS.BusinessEntity.Business
             ISupplierZoneDataManager dataManager = BEDataManagerFactory.GetDataManager<ISupplierZoneDataManager>();
             return dataManager.GetSupplierZones(supplierId, effectiveDate);
         }
-      
-        public int ReserveIDRange(int numberOfIDs)
+
+        public long ReserveIDRange(int numberOfIDs)
         {
             ISupplierZoneDataManager dataManager = BEDataManagerFactory.GetDataManager<ISupplierZoneDataManager>();
             return dataManager.ReserveIDRange(numberOfIDs);
